edit 4

+++ hapus-tugas.php
@@ -2,19 +2,32 @@
 include 'koneksi.php';
 
 if (isset($_GET['id_tugas'])) {
-    $id = $_GET['id_tugas'];
+    $id_tugas = $_GET['id_tugas'];
 
-    $sql = "DELETE FROM tugas WHERE id_tugas = :id_tugas";
-    $stmt = $conn->prepare($sql);
-    $stmt->bindParam(':id_tugas', $id);
+    // Ambil id_mapel dari tugas yang akan dihapus
+    $stmt = $conn->prepare("SELECT id_mapel FROM tugas WHERE id_tugas = :id_tugas");
+    $stmt->bindParam(':id_tugas', $id_tugas);
+    $stmt->execute();
+    $result = $stmt->fetch(PDO::FETCH_ASSOC);
 
-    if ($stmt->execute()) {
-        // Redirect ke halaman daftar film
-        header("Location: index-mapel.php?id=$id_mapel");
-        exit();
+    if ($result) {
+        $id_mapel = $result['id_mapel'];
+
+        // Hapus tugas
+        $sql = "DELETE FROM tugas WHERE id_tugas = :id_tugas";
+        $stmt = $conn->prepare($sql);
+        $stmt->bindParam(':id_tugas', $id_tugas);
+
+        if ($stmt->execute()) {
+            // Redirect ke halaman daftar mapel dengan id_mapel
+            header("Location: index-mapel.php?id=$id_mapel");
+            exit();
+        } else {
+            echo "Gagal menghapus data.";
+        }
     } else {
-        echo "Gagal menghapus data.";
+        echo "Data tugas tidak ditemukan.";
     }
 } else {
-    echo "ID film tidak ditemukan.";
+    echo "ID tugas tidak ditemukan.";
 }
